src/test/java/br/com/alura/adopet/api/validacoes/ValidacaoTutorComLimiteDeAdocoesTest.java
package br.com.alura.adopet.api.validacoes;

import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.BDDMockito.given;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import br.com.alura.adopet.api.dto.SolicitacaoAdocaoDto;
import br.com.alura.adopet.api.exception.ValidacaoException;
import br.com.alura.adopet.api.model.StatusAdocao;
import br.com.alura.adopet.api.repository.AdocaoRepository;

@ExtendWith(MockitoExtension.class)
class ValidacaoTutorComLimiteDeAdocoesTest {

	@InjectMocks
	private ValidacaoTutorComLimiteDeAdocoes validador;

	@Mock
	private AdocaoRepository adocaoRepository;

	@Mock
	private SolicitacaoAdocaoDto dto;

	@Test
	void naoDeveriaPermitirSolicitacaoDeAdocaoTutorAtingiuLimiteDe5Adocoes() {
		// ARRANGE
		given(adocaoRepository.countByTutorIdAndStatus(
				dto.idTutor(), 
				StatusAdocao.APROVADO)
	    ).willReturn(5);

		// ACT + ASSERT
		assertThrows(ValidacaoException.class, () -> validador.validar(dto));
	}

	@Test
	void deveriaPermitirSolicitacaoDeAdocaoTutorNaoAtingiuLimiteDe5Adocoes() {
		// ARRANGE
		given(adocaoRepository.countByTutorIdAndStatus(
				dto.idTutor(), 
				StatusAdocao.APROVADO)
		).willReturn(4);

		// ACT + ASSERT
		assertDoesNotThrow(() -> validador.validar(dto));
	}

}
